+++ CHANGELOG.md
@@ -11,13 +11,14 @@ All notable, unreleased changes to this project will be documented in this file.
 
 - Add `orderNoteAdd` and `orderNoteUpdate` mutations and deprecate `orderAddNote` mutation - #12434 by @pawelzar
 - Deprecate `Order.trackingClientId` field - #13146 by @SzymJ
+- Add error handling for invalid input fields in the `AccountAddressCreate` mutation. - #13086 by @FremahA
 
 ### Saleor Apps
 - Introduce `Saleor-Schema-Version` HTTP header in app manifest fetching and app installation handshake requests. - #13075 by @przlada
 
 ### Other changes
 - Expand metric units to support more types of products. - #13043 by @FremahA
-- Add error handling for invalid input fields in the `AccountAddressCreate` mutation. - #13086 by @FremahA
+
 - Fix Error Cannot return null for non-nullable field Webhook.name. - #12989 by @cyborg7898
 - Remove unused `django-versatileimagefield` package - #13148 by @SzymJ
 - Remove unused `google-measurement-protocol` package - #13146 by @SzymJ